+++ src/Interfaces/Item/ItemPropertiesBackpackInterface.php
@@ -8,10 +8,10 @@ interface ItemPropertiesBackpackInterface
 {
     public function getCapacity(): int;
     public function setCapacity(int $capacity): ItemPropertiesBackpackInterface;
-    public function getSpeedPenalty(): int;
-    public function setSpeedPenalty(int $speedPenalty): ItemPropertiesBackpackInterface;
-    public function getTurnPenalty(): int;
-    public function setTurnPenalty(int $turnPenalty): ItemPropertiesBackpackInterface;
+    public function getSpeedPenalty(): float;
+    public function setSpeedPenalty(float $speedPenalty): ItemPropertiesBackpackInterface;
+    public function getTurnPenalty(): float;
+    public function setTurnPenalty(float $turnPenalty): ItemPropertiesBackpackInterface;
     public function getErgoPenalty(): int;
     public function setErgoPenalty(int $ergoPenalty): ItemPropertiesBackpackInterface;
 }
