+++ src/Interfaces/Item/ItemPropertiesNightVisionInterface.php
@@ -4,12 +4,12 @@ namespace App\Interfaces\Item;
 
 interface ItemPropertiesNightVisionInterface
 {
-    public function getIntensity(): int;
-    public function setIntensity(int $intensity): ItemPropertiesNightVisionInterface;
-    public function getNoiseIntensity(): int;
-    public function setNoiseIntensity(int $noiseIntensity): ItemPropertiesNightVisionInterface;
-    public function getNoiseScale(): int;
-    public function setNoiseScale(int $noiseScale): ItemPropertiesNightVisionInterface;
-    public function getDiffuseIntensity(): int;
-    public function setDiffuseIntensity(int $diffuseIntensity): ItemPropertiesNightVisionInterface;
+    public function getIntensity(): float;
+    public function setIntensity(float $intensity): ItemPropertiesNightVisionInterface;
+    public function getNoiseIntensity(): float;
+    public function setNoiseIntensity(float $noiseIntensity): ItemPropertiesNightVisionInterface;
+    public function getNoiseScale(): float;
+    public function setNoiseScale(float $noiseScale): ItemPropertiesNightVisionInterface;
+    public function getDiffuseIntensity(): float;
+    public function setDiffuseIntensity(float $diffuseIntensity): ItemPropertiesNightVisionInterface;
 }
